+++ src/Functions/GatherUtils/Sources.h
@@ -328,7 +328,7 @@ private:
         ssize_t nested_row = replication_indexes.getIndexAt(logical_row);
         row_num = nested_row;
         /// `offsets[-1]` is a guaranteed zero (`PaddedPODArray` left padding), same as `ColumnArray::offsetAt`.
-        prev_offset = offsets[nested_row - 1];
+        prev_offset = offsets[static_cast<ssize_t>(nested_row) - 1];
     }
 };
