+++ src/Functions/overlay.cpp
@@ -376,8 +376,8 @@ private:
         size_t res_offset = 0;
         for (size_t i = 0; i < rows; ++i)
         {
-            size_t input_offset = input_offsets[i - 1];
-            size_t input_bytes = input_offsets[i] - input_offsets[i - 1];
+            size_t input_offset = input_offsets[static_cast<ssize_t>(i) - 1];
+            size_t input_bytes = input_offsets[i] - input_offsets[static_cast<ssize_t>(i) - 1];
             size_t input_size = getSliceSize<is_utf8_>(&input_data[input_offset], input_bytes);
 
             if constexpr (offset_is_const)
@@ -492,8 +492,8 @@ private:
         size_t res_offset = 0;
         for (size_t i = 0; i < rows; ++i)
         {
-            size_t replace_offset = replace_offsets[i - 1];
-            size_t replace_bytes = replace_offsets[i] - replace_offsets[i - 1];
+            size_t replace_offset = replace_offsets[static_cast<ssize_t>(i) - 1];
+            size_t replace_bytes = replace_offsets[i] - replace_offsets[static_cast<ssize_t>(i) - 1];
             size_t replace_size = getSliceSize<is_utf8_>(&replace_data[replace_offset], replace_bytes);
 
             if constexpr (!offset_is_const)
@@ -609,12 +609,12 @@ private:
         size_t res_offset = 0;
         for (size_t i = 0; i < rows; ++i)
         {
-            size_t input_offset = input_offsets[i - 1];
-            size_t input_bytes = input_offsets[i] - input_offsets[i - 1];
+            size_t input_offset = input_offsets[static_cast<ssize_t>(i) - 1];
+            size_t input_bytes = input_offsets[i] - input_offsets[static_cast<ssize_t>(i) - 1];
             size_t input_size = getSliceSize<is_utf8_>(&input_data[input_offset], input_bytes);
 
-            size_t replace_offset = replace_offsets[i - 1];
-            size_t replace_bytes = replace_offsets[i] - replace_offsets[i - 1];
+            size_t replace_offset = replace_offsets[static_cast<ssize_t>(i) - 1];
+            size_t replace_bytes = replace_offsets[i] - replace_offsets[static_cast<ssize_t>(i) - 1];
             size_t replace_size = getSliceSize<is_utf8_>(&replace_data[replace_offset], replace_bytes);
 
             if constexpr (offset_is_const)
